+++ packages/eslint-config/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.0
 
 - feat: add typescript strict rules
+- feat: add `@typescript-eslint/indent` rule to avoid `cannot read property 'loc' of undefined` problem. Ref: <https://github.com/eslint/eslint/issues/13956>
 
 ## 1.0.2